feat: refactor import path for RegisterForm component to align with new directory structure
